+++ tests/test_benchmarks.py
@@ -39,7 +39,7 @@ def test_findall(benchmark):
     def finding():
         random.seed(999)
         i = int.to_bytes(random.getrandbits(20000000), 20000000 // 8, 'big')
-        s = bitstring.BitArray(bytes=i)
+        s = bitstring.BitArray.from_bytes(i)
         for ss in ['0b11010010101', '0xabcdef1234, 0b000101111010101010011010100100101010101', '0x4321']:
             x = len(list(s.findall(ss)))
         return x
@@ -50,7 +50,7 @@ def test_repeated_reading(benchmark):
     def repeating_reading():
         random.seed(1414)
         i = int.to_bytes(random.getrandbits(100000*8), 100000, 'big')
-        s = bitstring.Reader(bitstring.Bits(bytes=i))
+        s = bitstring.Reader(bitstring.Bits.from_bytes(i))
         for _ in range(800000 // 40):
             _ = s.readlist('uint:4, float:32, bool, bool, bool, bool')
     benchmark(repeating_reading)